% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/economics_reduced.R
\docType{data}
\name{economics_reduced}
\alias{economics_reduced}
\title{Italian Economics Faculties Dataset (2009/2010)}
\format{
A data frame with 55 observations on the following 14 variables:
\describe{
\item{\code{University_Type}}{Type of university: Private or Public (Factor)}
\item{\code{P1}}{Rate of persistence between the first and second year (Numeric)}
\item{\code{P2}}{Achieved credits (Numeric)}
\item{\code{P3A}}{Rate of regular students in 3-year bachelor programmes (Numeric)}
\item{\code{P3B}}{Rate of regular students in 2-year master programmes (Numeric)}
\item{\code{P4A}}{Rate of regular graduates in 3-year bachelor programmes (Numeric)}
\item{\code{P4B}}{Rate of regular graduates in 2-year master programmes (Numeric)}
}
}
\source{
\insertRef{ICR}{msds}
\url{https://search.r-project.org/CRAN/refmans/datasetsICR/html/Economics.html}
}
\usage{
economics_reduced
}
\description{
A dataset containing performance indicators for 55 Italian Economics faculties
during the 2009/2010 academic year, used for evaluating productivity and
teaching quality \insertCite{ICR}{msds} .
}
\details{
The dataset includes 6 productivity indicators (P1-P4B) to evaluate the performance of 48 Italian Economics
faculties working at public universities in the 2009/2010 academic year.  This is a reduced data set found in \insertCite{ICR}{msds}  for educational purposes.
}
\keyword{datasets}
