% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UserFunctions.R
\name{pca.scree}
\alias{pca.scree}
\title{Scree Plots Derived From PCA}
\usage{
pca.scree(x, eig.vals = NULL, ref = 0.8, label.size = 11)
}
\arguments{
\item{x}{An object containing a \code{sdev} component, such as that returned by \link[stats:princomp]{stats::princomp} or \link{stats:prcomp} .}

\item{eig.vals}{Vector of eigenvalues from a PCA decomposition.  Alternative to \code{x}.}

\item{ref}{Horizontal reference line for the cumulative proportion plot.}

\item{label.size}{Text size for x and y axis labels.}
}
\value{
A 1x3 panel of scree plots using the eigenvalues, proportion of variance explained, and cumulative proportion of variance explained.
}
\description{
Produces scree plots from user provided eigenvalues or an object created by \link[stats:princomp]{stats::princomp} or \link{stats:prcomp}.
}
\details{
The user may specify eigenvalues for plotting rather than an object created by one of the common PCA functions. User should specify \code{x=NULL} when
using the alternative option.
}
\examples{
faculty.pca<-prcomp(economics_reduced)
pca.scree(facult.pca,ref=.9,label.size=14)




}
