% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UserFunctions.R
\name{lc.trans}
\alias{lc.trans}
\title{Linear Combination of Variables}
\usage{
lc.trans(y, C)
}
\arguments{
\item{y}{A data frame or numeric matrix.}

\item{C}{Matrix of weights.}
}
\value{
A data frame of the linear combinations:
\item{LRT.Layers}{A data frame with the Lilkelihood Ratio Test Statistic for each layer of the tree.}
\item{Overall}{The overall Lilkelihood Ratio Test Statistics and p-value.}
}
\description{
Performs linear combinations using a data matrix y.
}
\details{
Generates a new data matrix that are linear combinations of the data matrix y.
}
\examples{
y<-data.frame(Y1=rnorm(10),Y2=rnorm(10),Y3=rnorm(10))
C1<-data.frame(Z1=c(1,1,1),Z2=c(1,0,0),Z3=c(5,2,4))
lc.trans(y,C)



}
